aider: Created a BeforeAfterSlider component that allows users to move a slider to reveal before and after images.
# Context:
USER: create a before and after slider component this should allow people to move as slider which will show a before and after image on either side and the slider determined how much of which images is shown
ASSISTANT: To create a before and after slider component, we will need to create a new file in your components directory. This component will use JavaScript to manipulate the CSS properties of two images to create the effect of a slider revealing one image over another.

Here are the steps we will follow:

1. Create a new file in the components directory named `BeforeAfterSlider.tsx`.
2. In this file, we will define a new React component named `BeforeAfterSlider`.
3. This component will render two images, one on top of the other, with the top image's width controlled by the position of a slider.
4. We will use the `useState` hook to keep track of the slider's position.
5. We will use the `useEffect` hook to update the width of the top image whenever the slider's position changes.

Here is the code for the new component:

```tsx
sanity/components/BeforeAfterSlider.tsx
<<<<<<< ORIGINAL
=======
import React, { useState, useEffect } from 'react';

const BeforeAfterSlider = ({ beforeImage, afterImage }) => {
  const [sliderPosition, setSliderPosition] = useState(50);

  useEffect(() => {
    const topImage = document.getElementById('top-image');
    if (topImage) {
      topImage.style.width = `${sliderPosition}%`;
    }
  }, [sliderPosition]);

  return (
    <div className="relative">
      <img src={beforeImage} alt="Before" className="w-full" />
      <img
        id="top-image"
        src={afterImage}
        alt="After"
        className="absolute top-0"
      />
      <input
        type="range"
        min="0"
        max="100"
        value={sliderPosition}
        onChange={(e) => setSliderPosition(e.target.value)}
        className="absolute bottom-0 w-full"
      />
    </div>
  );
};

export default BeforeAfterSlider;
>>>>>>> UPDATED
```

This component takes two props: `beforeImage` and `afterImage`, which are the URLs of the images to display. The `BeforeAfterSlider` component can be used in any other component by importing it and passing the appropriate props. For example:

```tsx
import BeforeAfterSlider from './BeforeAfterSlider';

// ...

<BeforeAfterSlider beforeImage="/path/to/before.jpg" afterImage="/path/to/after.jpg" />
```

Please make sure to replace `"/path/to/before.jpg"` and `"/path/to/after.jpg"` with the actual paths to your images.